refactor: Order 조회에 Optional 추가
